Organize search layout components
