+++ src/pages/search/dashboard/landscape/HumanFootprint.tsx
@@ -1,6 +1,6 @@
 import AddIcon from "@mui/icons-material/Add";
 
-import Accordion from "pages/search/Accordion";
+import Accordion from "pages/search/layout/Accordion";
 import CurrentFootprint from "pages/search/dashboard/landscape/humanFootprint/CurrentFootprint";
 import PersistenceFooprint from "pages/search/dashboard/landscape/humanFootprint/PersistenceFootprint";
 import TimelineFootprint from "pages/search/dashboard/landscape/humanFootprint/TimelineFootprint";
