+++ src/pages/search/dashboard/landscape/PAConnectivity.tsx
@@ -1,4 +1,4 @@
-import Accordion from "pages/search/Accordion";
+import Accordion from "pages/search/layout/Accordion";
 import CurrentPAConnectivity from "pages/search/dashboard/landscape/connectivity/CurrentPAConnectivity";
 import TimelinePAConnectivity from "pages/search/dashboard/landscape/connectivity/TimelinePAConnectivity";
 import CurrentSEPAConnectivity from "pages/search/dashboard/landscape/connectivity/CurrentSEPAConnectivity";
