+++ src/pages/search/layout/Selector.tsx
@@ -2,7 +2,7 @@ import { useEffect, useState } from "react";
 
 import EditIcon from "@mui/icons-material/Edit";
 
-import Accordion from "pages/search/Accordion";
+import Accordion from "pages/search/layout/Accordion";
 import isFlagEnabled from "core/utils/isFlagEnabled";
 import {
   ErrorMessage,
